Wired up so now the server runs and shows data for a zip code. Still need to add a frontend, host in azure, send email with links to code and azure. Potentially put in some logging, or some more edge case tests... nah... just comment and be on your way.

+++ ZipcodeInfo/ApiClients/GoogleMapsApiClient.cs
@@ -18,7 +18,7 @@ namespace ZipcodeInfo.ApiClients
         }
 
 
-        public async Task<ApiResponse<string>> GetTimeZoneAsync(double lon, double lat, long timeStamp)
+        public async Task<ApiResponse<string>> GetTimeZoneAsync(double lon, double lat, double timeStamp)
         {
             var response = new ApiResponse<string>();
